Fixup: Endpoint build issues
