+++ src/Moryx.Material.Endpoints/Model/MaterialContainerTypeModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 // Licensed under the Apache License, Version 2.0
 
-namespace Moryx.Material.Endpoints;
+namespace Moryx.Material.Endpoints.Model;
 
 public class MaterialContainerTypeModel
 {
