+++ src/Moryx.Material.Endpoints/Model/PreAdvideModel.cs
@@ -1,9 +1,9 @@
-﻿// Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
+// Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 // Licensed under the Apache License, Version 2.0
 
 using System.Runtime.Serialization;
 
-namespace Moryx.Material.Endpoints;
+namespace Moryx.Material.Endpoints.Model;
 
 [DataContract]
 public enum PreAdviceDepartureReasonModel
