+++ src/Moryx.Material.Endpoints/Moryx.Material.Endpoints.csproj
@@ -4,7 +4,7 @@
     <TargetFramework>net10.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <Description>REST Api in order to interact with the resource management</Description>
-    <PackageTags>MORYX;IIoT;IoT;Resource</PackageTags>
+    <PackageTags>MORYX;IIoT;IoT;Material</PackageTags>
 	  <IsPackable>true</IsPackable>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
@@ -33,6 +33,10 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Strings.resx</DependentUpon>
     </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="C:\Repositories\00 Products\MORYX-Framework\.build\\moryx-logo.png" Link="Model\moryx-logo.png" />
   </ItemGroup>
 
 </Project>
